[Bug] Execution bug caused by sql fragments split by flinksql not sorted in order
